pyproject.toml
[tool.poetry]
name = "breach-bot"
version = "0.3.0"
description = ""
authors = ["chahan69 <chahan69@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.*"
environ-config = "^21.2.0"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
pyfiglet = "^0.8.post1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
